DevOps (Docker)/api/requirements.txt

blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
python-dotenv==1.0.0
Werkzeug==3.1.3
